https://philippstracker.com/divi-popup/
Turns a section into a popup. Important: Every popup needs an ID, otherwise it can not be called.

Add your own "Close" feature in your popup. Simply insert a link, image or other element with the class `close` in the respective popup.

In addition, add the class `is-modal` to a pop-up section and the pop-up will not be closed when the user clicks on the gray background. In that case you need a `close` element to close the popup.

Displays the popup when the visitor tries to leave the page. By the way: This works by observing the movement of the mouse. That's why it works only on desktop computers with a mouse, and not on devices with touch screen such as mobile phones.

Easy
Just add a CSS class to any section and you have a popup.

Unlimited
Create as many popups as you need. Or want. It is fun!

Visual Builder
Turn any section in the visual builder directly into a popup!

Flexible
Just show your popups by clicking. For example as a link or button.

Responsive
Do not worry about mobile optimizations. Your popups are completely responsive.

Free
Divi Popups is and remains free, fully compatible with GPL compatible license.

Fast
Each line of code is optimized to load your page as fast as possible. Fully compatible with all caching plugins.

Developers love it
A stable JavaScript API and customization options via WordPress Filter make the plugin even more versatile!

To install
Search, install and activate the Divi Popup plugin via the Admin Dashboard.

Define popup
In the Visual Builder add the CSS class "popup" and the ID "example-1" to a section

Call popup
Just add a button and enter the link destination "# example-1"

Finished
Upload your page without the Visual Builder and you can use your first popup!

See example-1

The filter is applied to an array and allows you to change JavaScript defaults. Most options should be ignored. Sometimes it may happen that another plugin or child theme is incompatible with your pop-ups. Then you might find a solution here.